requirements.txt
langchain-groq
langchain
chromadb
langchain_community
beautifulsoup4
pandas
streamlit
python-dotenv